code/server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sifter-server"
version = "0.1.0"
description = "AI-powered document extraction engine — self-hostable server"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
authors = [{ name = "Sifter Contributors" }]
keywords = ["document", "extraction", "ai", "mongodb", "llm", "ocr"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "motor>=3.4.0",
    "pymongo>=4.7.0",
    "litellm>=1.40.0",
    "pymupdf>=1.24.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-multipart>=0.0.9",
    "structlog>=24.1.0",
    "aiofiles>=23.2.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.0.0",
    "slowapi>=0.1.9",
    "email-validator>=2.1.0",
    "google-auth>=2.29.0",
]

[dependency-groups]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-mock>=3.14.0",
    "respx>=0.21.0",
    "fpdf2>=2.8.7",
    "sifter-ai",
]

[project.scripts]
sifter-server = "sifter.server:run"

[project.urls]
Homepage = "https://github.com/bfortunato/sifter"
Repository = "https://github.com/bfortunato/sifter"

[tool.hatch.build.targets.wheel]
packages = ["sifter"]

[tool.uv]
package = true

[project.optional-dependencies]
mcp = ["sifter-mcp>=0.1.0"]

[tool.uv.sources]
sifter-ai = { path = "../sdk", editable = true }
sifter-mcp = { path = "../mcp", editable = true }

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
